wrapper: Add some Ex functions to LowLevelApi
* C_EX_InitToken;
* C_EX_SetLicense;
* C_EX_GetLicense;
* C_EX_SetLocalPIN;

Refer #4

+++ pkcs11wrapper/lib.pkcs11wrapper/src/main/java/ru/rutoken/pkcs11wrapper/lowlevel/jna/Pkcs11JnaLowLevelApi.java
@@ -116,7 +116,7 @@ public class Pkcs11JnaLowLevelApi implements IPkcs11LowLevelApi {
     }
 
     protected static void assign(MutableLong value, NativeLongByReference nativeValue) {
-        value.value = nativeValue.getValue().longValue();
+        value.value = unsigned(nativeValue.getValue());
     }
 
     protected static NativeLong length(byte @Nullable [] data) {
